+++ src/platform/ui_engine/patterns/index.tsx
@@ -6,6 +6,7 @@ export * from "./creatable-search";
 export * from "./draft-dialog";
 export * from "./file-drop-zone";
 export * from "./hooks";
+export * from "./image-workspace";
 export * from "./inline-edit";
 export * from "./pagination";
 export * from "./simple-text-editor";